+++ jannovar-stats/src/main/java/de/charite/compbio/jannovar/stats/facade/Genotype.java
@@ -1,0 +1,15 @@
+package de.charite.compbio.jannovar.stats.facade;
+
+/**
+ * Alternative genotype types for statistics
+ * 
+ * @author <a href="mailto:manuel.holtgrewe@bihealth.de">Manuel Holtgrewe</a>
+ */
+public enum Genotype {
+	/** Heterozygous alt */
+	HET,
+	/** Homozygous alt */
+	HOM_ALT,
+	/** Hemizygous alt */
+	HEM_ALT;
+}