+++ jannovar-stats/src/main/java/de/charite/compbio/jannovar/stats/facade/TsTv.java
@@ -1,0 +1,14 @@
+package de.charite.compbio.jannovar.stats.facade;
+
+/**
+ * Enumeration for the transition/transversion classification of SNVs
+ *
+ *
+ * @author <a href="mailto:manuel.holtgrewe@bihealth.de">Manuel Holtgrewe</a>
+ */
+public enum TsTv {
+	/** Transition */
+	TS,
+	/** Transversion */
+	TV;
+}